Match gm_801A9DD0 (gmopening) — fix arg2 type int->u16
gm_801A9DD0's 3rd parameter was mistyped as `int` (in both the header decl
and the .c def, same as upstream). That forced an int->float u16-mask magic
conversion, whose 8-byte double scratch reserved an unused 4-byte stack home +
4 bytes of alignment -> frame 0x98 vs the target 0x90, walling the function at
99.60% (instruction-identical, only the frame differed). The real signature is
`u16 arg2` (all callers pass u16-fitting 0x280/0x1E0); with it, no conversion,
no scratch, no home -> frame 0x90 -> byte match=true (100%).

Found by the frame-layer tooling deep-dive: inspect frame-reservations + the
frame-aware diagnose pinpointed the unused home as alignment growth from the
int->float magic scratch created by the int arg2 u16-mask, which pointed
straight at the parameter type.

+++ src/melee/gm/gmopening.c
@@ -30,7 +30,7 @@ static PerfLabelLine gm_80480B38[4];
 static const Vec3 gm_803B7D68 = { 0.0f, 0.0f, 1.0f };
 static const Vec3 gm_803B7D74 = { 0.0f, 0.0f, 0.0f };
 
-void gm_801A9DD0(HSD_GObj* arg0, u16 arg1, int arg2, int arg3, int arg4)
+void gm_801A9DD0(HSD_GObj* arg0, u16 arg1, u16 arg2, int arg3, int arg4)
 {
     HSD_CObj* cobj;
     f32 roll = 0.0f;
